feat: don't let locked threads be deleted

+++ api/models/forum.js
@@ -126,6 +126,8 @@ async function deleteThread(id) {
 
   if (threadIndex === -1) throw new Error('Fil de discussion introuvable');
 
+  if (threads[threadIndex].locked) throw new Error('Le fil de discussion est verrouillé');
+
   await db.delete(`/forum[${threadIndex}]`);
 }
